Refactored event namespace.
