+++ ConnectionStringTest/ConnectionStringTest.csproj
@@ -47,9 +47,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Utils\ConnectionStringTester.cs" />
-    <Compile Include="Event\IUiEventHandler.cs" />
-    <Compile Include="Event\TestButtonClickHandler.cs" />
-    <Compile Include="Event\UiEvent.cs" />
+    <Compile Include="EventHandling\Handler.cs" />
+    <Compile Include="EventHandling\TestButtonClickHandler.cs" />
+    <Compile Include="EventHandling\Event.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestResponse.cs" />
